+++ src/services/topic/pglite.test.ts
@@ -5,7 +5,7 @@ import { clientDB, initializeDB } from '@/database/client/db';
 import { sessions, topics, users } from '@/database/schemas';
 import { ChatTopic } from '@/types/topic';
 
-import { ClientService } from './pglite';
+import { ClientService } from './client';
 
 // Mock data
 const userId = 'topic-user-test';
